added test for execute core to sync
